Asg_4_part_2_files/requirements.txt
pandas
numpy
faker
uvicorn
fastapi
requests
snowflake-sqlalchemy
openai
groq
streamlit
snowflake-snowpark-python[pandas]